services/user-policy-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
email-validator==2.2.0

passlib[bcrypt]
python-jose[cryptography]
python-multipart
